added another slightly more complex test

+++ src/test/java/com/github/fge/largetext/MutableLargeTextFileFixture.java
@@ -12,6 +12,7 @@ import java.nio.file.Paths;
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.testng.FileAssert.fail;
 
 /**
  * Created by Geoff on 3/26/2016.
@@ -83,6 +84,33 @@ public class MutableLargeTextFileFixture {
         );
     }
 
+    @Test
+    public void when_inserting_content_at_the_front_middle_and_back_should_appear() throws IOException {
+        //setup
+        MutableLargeTextFile mutable = makeMutableText(SimpleThreeLineUTF8Doc);
+        fileUnderTest = generateNewFileFor(SimpleThreeLineUTF8Doc);
+
+        //act
+        mutable.append("Front!\n", 0, 1);
+        mutable.append("\nMiddle!\n", 53, 54);
+        //uhh, 53 + "Front!\n".length() ?
+        //maybe use content.indexOf("456") for clarity?
+        fail("TODO update indexes");
+        mutable.append("Back!\n", 104, 105);
+        mutable.writeTo(fileUnderTest);
+
+        //assert
+        assertThat(fileUnderTest.toFile()).hasContent(
+                "Front!\n" +
+                "This is the first line, it only contains words\n" +
+                "1.23\n" +
+                "Middel!\n" +
+                "456\n" +
+                "This is the last line, it contains more words\n"
+//                "Back!\n"
+        );
+    }
+
     private MutableLargeTextFile makeMutableText(Path pathToPreloadedContent) throws IOException {
         LargeText largeText = LargeTextFactory.defaultFactory().load(pathToPreloadedContent);
         return new MutableLargeTextFile(largeText);
@@ -90,7 +118,7 @@ public class MutableLargeTextFileFixture {
 
     @AfterMethod
     public void delete_file_under_test() throws IOException {
-        if(fileUnderTest != null){
+        if(fileUnderTest != null && Files.exists(fileUnderTest)){
             Files.delete(fileUnderTest);
         }
     }
